# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AX Studio API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type AuditionRoleStatus = typeof AuditionRoleStatus[keyof typeof AuditionRoleStatus];


export const AuditionRoleStatus = {
  준비중: '준비중',
  접수중: '접수중',
  마감: '마감',
} as const;

export interface AuditionRole {
  id: number;
  roleName: string;
  ageRange: string;
  description: string;
  status: AuditionRoleStatus;
}

export type AuditionRoleInputStatus = typeof AuditionRoleInputStatus[keyof typeof AuditionRoleInputStatus];


export const AuditionRoleInputStatus = {
  준비중: '준비중',
  접수중: '접수중',
  마감: '마감',
} as const;

export interface AuditionRoleInput {
  /** @minLength 1 */
  roleName: string;
  /** @minLength 1 */
  ageRange: string;
  /** @minLength 1 */
  description: string;
  status: AuditionRoleInputStatus;
}

export interface AuditionApplicationInput {
  roleId: number;
  childName: string;
  birthYear: string;
  gender: string;
  guardianName: string;
  phone: string;
  /** @nullable */
  portfolio?: string | null;
  /** @nullable */
  memo?: string | null;
}

export interface AuditionApplication {
  id: number;
  roleId: number;
  childName: string;
  birthYear: string;
  gender: string;
  guardianName: string;
  phone: string;
  /** @nullable */
  portfolio?: string | null;
  /** @nullable */
  memo?: string | null;
  createdAt: string;
}

export type SponsorshipInquiryInputSponsorType = typeof SponsorshipInquiryInputSponsorType[keyof typeof SponsorshipInquiryInputSponsorType];


export const SponsorshipInquiryInputSponsorType = {
  장소협찬: '장소협찬',
  '기업(제품)협찬': '기업(제품)협찬',
  기타: '기타',
} as const;

export interface SponsorshipInquiryInput {
  organizationName: string;
  contactName: string;
  contactEmail: string;
  /** @nullable */
  contactPhone?: string | null;
  sponsorType: SponsorshipInquiryInputSponsorType;
  message: string;
}

export interface SponsorshipInquiry {
  id: number;
  organizationName: string;
  contactName: string;
  contactEmail: string;
  /** @nullable */
  contactPhone?: string | null;
  sponsorType: string;
  message: string;
  createdAt: string;
}

export interface NewsItem {
  id: number;
  date: string;
  category: string;
  title: string;
  summary: string;
  /** @nullable */
  badge: string | null;
  /** @nullable */
  link: string | null;
  isRecent: boolean;
}

export interface NewsItemInput {
  /** @minLength 1 */
  date: string;
  /** @minLength 1 */
  category: string;
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  summary: string;
  /** @nullable */
  badge?: string | null;
  /** @nullable */
  link?: string | null;
  isRecent: boolean;
}

export interface SiteStats {
  totalApplications: number;
  openRoles: number;
  totalInquiries: number;
}