# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AX Studio API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /news:
    get:
      operationId: listNewsItems
      tags: [news]
      summary: List news items
      responses:
        "200":
          description: List of news items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NewsItem"

    post:
      operationId: createNewsItem
      tags: [news]
      summary: Create a news item (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewsItemInput"
      responses:
        "201":
          description: News item created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NewsItem"
        "400":
          description: Validation error
  /news/{newsId}:
    put:
      operationId: updateNewsItem
      tags: [news]
      summary: Update a news item (admin)
      parameters:
        - name: newsId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewsItemInput"
      responses:
        "200":
          description: News item updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NewsItem"
        "400":
          description: Validation error
        "404":
          description: News item not found

    delete:
      operationId: deleteNewsItem
      tags: [news]
      summary: Delete a news item (admin)
      parameters:
        - name: newsId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: News item deleted
        "404":
          description: News item not found
...
components:
  schemas:
...
    NewsItem:
      type: object
      required: [id, date, category, title, summary, badge, link, isRecent]
      properties:
        id:
          type: integer
        date:
          type: string
        category:
          type: string
        title:
          type: string
        summary:
          type: string
        badge:
          type: ["string", "null"]
        link:
          type: ["string", "null"]
        isRecent:
          type: boolean
    NewsItemInput:
      type: object
      required: [date, category, title, summary, isRecent]
      properties:
        date:
          type: string
          minLength: 1
        category:
          type: string
          minLength: 1
        title:
          type: string
          minLength: 1
        summary:
          type: string
          minLength: 1
        badge:
          type: ["string", "null"]
        link:
          type: ["string", "null"]
        isRecent:
          type: boolean
...