# lib/api-zod/src/generated/types/newsItem.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AX Studio API specification
 * OpenAPI spec version: 0.1.0
 */

export interface NewsItem {
  id: number;
  date: string;
  category: string;
  title: string;
  summary: string;
  /** @nullable */
  badge: string | null;
  /** @nullable */
  link: string | null;
  isRecent: boolean;
}