# lib/api-zod/src/generated/types/newsItemInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AX Studio API specification
 * OpenAPI spec version: 0.1.0
 */

export interface NewsItemInput {
  /** @minLength 1 */
  date: string;
  /** @minLength 1 */
  category: string;
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  summary: string;
  /** @nullable */
  badge?: string | null;
  /** @nullable */
  link?: string | null;
  isRecent: boolean;
}